print_honey_import_info is moved into independet file
